pyproject.toml
[tool.poetry]
name = "actorshere"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2"
jwt = "^1.3.1"
psycopg2 = "^2.9.6"
dotenv = "^0.0.5"
python-dotenv = "^1.0.0"
djangorestframework = "^3.14.0"
djoser = "^2.2.0"
factory-boy = "^3.2.1"
django-cors-headers = "^3.14.0"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
pre-commit = "^3.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
